+++ src/hotspot/share/logging/logStream.hpp
@@ -32,8 +32,7 @@
 #include "utilities/align.hpp"
 #include "utilities/ostream.hpp"
 
-template<typename BackingLog>
-class LogStreamImpl : public outputStream {
+class LogStreamImplBase : public outputStream {
   friend class LogStreamTest_TestLineBufferAllocation_vm_Test;
   friend class LogStreamTest_TestLineBufferAllocationCap_vm_Test;
 
@@ -65,135 +64,36 @@ class LogStreamImpl : public outputStream {
     void reset();
   };
 
-private:
-  LineBuffer _current_line;
-
 protected:
+  LineBuffer _current_line;
+};
+
+template<typename BackingLog>
+class LogStreamImpl : public LogStreamImplBase {
+private:
   BackingLog _backing_log;
 
 public:
   explicit LogStreamImpl(BackingLog bl)
     : _backing_log(bl) {};
 
-  virtual ~LogStreamImpl() {
-    if (!_current_line.is_empty()) {
-      _backing_log.print("%s", _current_line.buffer());
-      _current_line.reset();
-    }
-  }
+  ~LogStreamImpl() override;
 
   bool is_enabled() {
     return _backing_log.is_enabled();
   }
 
-  void write(const char* s, size_t len) override {
-    if (len > 0 && s[len - 1] == '\n') {
-      _current_line.append(s, len - 1); // omit the newline.
-      _backing_log.print("%s", _current_line.buffer());
-      _current_line.reset();
-    } else {
-      _current_line.append(s, len);
-    }
-    update_position(s, len);
-  }
+  void write(const char* s, size_t len) override;
 };
-
-template<typename T>
-LogStreamImpl<T>::LineBuffer::LineBuffer()
-  : _buf(_smallbuf), _cap(sizeof(_smallbuf)), _pos(0) {
-  _buf[0] = '\0';
-}
-
-template<typename T>
-LogStreamImpl<T>::LineBuffer::~LineBuffer() {
-  assert(_pos == 0, "still outstanding bytes in the line buffer");
-  if (_buf != _smallbuf) {
-    os::free(_buf);
-  }
-}
-
-// try_ensure_cap tries to enlarge the capacity of the internal buffer
-// to the given atleast value. May fail if either OOM happens or atleast
-// is larger than a reasonable max of 1 M.
-// Returns whether the capacity is at least atleast bytes.
-template<typename T>
-bool LogStreamImpl<T>::LineBuffer::try_ensure_cap(size_t atleast) {
-  // Cap out at a reasonable max to prevent runaway leaks.
-  const size_t reasonable_max = 1 * M;
-
-  assert(_cap >= sizeof(_smallbuf), "sanity");
-  assert(_cap <= reasonable_max, "sanity");
-
-  if (_cap >= atleast) {
-    return true;
-  }
-  if (_cap == reasonable_max) {
-    return false;
-  }
-
-  const size_t additional_expansion = 256;
-  size_t newcap = align_up(atleast + additional_expansion, additional_expansion);
-
-  if (newcap > reasonable_max) {
-    log_info(logging)("Suspiciously long log line: \"%.100s%s",
-                      _buf, (_pos >= 100 ? "..." : ""));
-    newcap = reasonable_max;
-  }
-
-  char* const newbuf = (char*)os::malloc(newcap, mtLogging);
-  if (newbuf == NULL) { // OOM. Leave object unchanged.
-    return false;
-  }
-  if (_pos > 0) { // preserve old content
-    memcpy(newbuf, _buf, _pos + 1); // ..including trailing zero
-  }
-  if (_buf != _smallbuf) {
-    os::free(_buf);
-  }
-  _buf = newbuf;
-  _cap = newcap;
-
-  return _cap >= atleast;
-}
-
-template<typename T>
-void LogStreamImpl<T>::LineBuffer::append(const char* s, size_t len) {
-  assert(_buf[_pos] == '\0', "sanity");
-  assert(_pos < _cap, "sanity");
-  const size_t minimum_capacity_needed = _pos + len + 1;
-  bool has_capacity = try_ensure_cap(minimum_capacity_needed);
-  // try_ensure_cap may not have enlarged the capacity to the full requested
-  // extent or may have not worked at all. In that case, just gracefully work
-  // with what we have already; just truncate if necessary.
-  if (!has_capacity) {
-    len = _cap - _pos - 1;
-    if (len == 0) {
-      return;
-    }
-  }
-  memcpy(_buf + _pos, s, len);
-  _pos += len;
-  _buf[_pos] = '\0';
-}
-
-template<typename T>
-void LogStreamImpl<T>::LineBuffer::reset() {
-  _pos = 0;
-  _buf[_pos] = '\0';
-}
 
 class LogStream : public LogStreamImpl<LogTargetHandle> {
   // see test/hotspot/gtest/logging/test_logStream.cpp
   friend class LogStreamTest_TestLineBufferAllocation_vm_Test;
   friend class LogStreamTest_TestLineBufferAllocationCap_vm_Test;
 
-  // No heap allocation of LogStream.
-  static void* operator new   (size_t) = delete;
-  static void* operator new[] (size_t) = delete;
   NONCOPYABLE(LogStream);
 
 public:
-  virtual ~LogStream() {};
   // Constructor to support creation from a LogTarget instance.
   //
   // LogTarget(Debug, gc) log;
@@ -234,7 +134,6 @@ public:
 template<LogLevelType level, LogTagType T0, LogTagType T1, LogTagType T2, LogTagType T3,
          LogTagType T4, LogTagType GuardTag>
 class LogStreamTemplate : public LogStream {
-
 public:
   LogStreamTemplate()
     : LogStream((LogTargetImpl<level, T0, T1, T2, T3, T4, GuardTag>*)NULL) {}
@@ -247,9 +146,11 @@ class LogMessageHandle {
 public:
   LogMessageHandle(const LogLevelType level, LogMessageImpl& lm)
     : _level(level), _lm(lm) {}
+
   bool is_enabled() {
     return _lm.is_level(_level);
   }
+
   void print(const char* fmt, ...) ATTRIBUTE_PRINTF(2, 3) {
     va_list args;
     va_start(args, fmt);
